long to int

+++ android/src/main/kotlin/com/github/jeremaihloo/fluent_aliyun_oss/AliyunOssClient.kt
@@ -67,7 +67,7 @@ class AliyunOssClient(context: Context?, private val clientConfig: AliyunOssClie
         return objectName
     }
 
-    fun signUrl(bucketName: String, objectKey: String, expiredTimeInSeconds: Long): String {
+    fun signUrl(bucketName: String, objectKey: String, expiredTimeInSeconds: Int): String {
         return ossClient.presignConstrainedObjectURL(bucketName, objectKey, expiredTimeInSeconds)
     }
